refactor: merge bottom & convert
